requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
chromadb==0.5.5
python-dotenv==1.0.1
pydantic==2.8.2
pypdf==5.0.1
markdownify==0.12.1
numpy==1.26.4
sentence-transformers==2.7.0
requests==2.32.3
